+++ nodejs/declarationHandler.js
@@ -46,7 +46,8 @@ const CLASSES_OF_TRUTH = [
     'ConfigSync',
     'DeviceGroup',
     'FailoverUnicast',
-    'Analytics'
+    'Analytics',
+    'ManagementRoute'
 ];
 
 /**