+++ nodejs/networkHandler.js
@@ -77,7 +77,7 @@ class NetworkHandler {
 function handleVlan() {
     return new Promise((resolve, reject) => {
         const promises = [];
-        forEach(this.declaration, 'VLAN', (tenant, vlan) => {
+        doUtil.forEach(this.declaration, 'VLAN', (tenant, vlan) => {
             if (vlan && vlan.name) {
                 const vlanInterfaces = vlan && Array.isArray(vlan.interfaces) ? vlan.interfaces.slice() : [];
                 const interfaces = [];
@@ -133,7 +133,7 @@ function handleSelfIp() {
     return new Promise((resolve, reject) => {
         const nonFloatingBodies = [];
         const floatingBodies = [];
-        forEach(this.declaration, 'SelfIp', (tenant, selfIp) => {
+        doUtil.forEach(this.declaration, 'SelfIp', (tenant, selfIp) => {
             if (selfIp && selfIp.name) {
                 let vlan;
 
@@ -244,7 +244,7 @@ function handleSelfIp() {
 function handleRoute() {
     return new Promise((resolve, reject) => {
         const promises = [];
-        forEach(this.declaration, 'Route', (tenant, route) => {
+        doUtil.forEach(this.declaration, 'Route', (tenant, route) => {
             if (route && route.name) {
                 const routeBody = {
                     name: route.name,
@@ -268,39 +268,6 @@ function handleRoute() {
                 logger.severe(`Error creating routes: ${err.message}`);
                 reject(err);
             });
-    });
-}
-
-/**
- * Iterates over the tenants in a parsed declaration
- *
- * At this point, Declarative Onboarding only supports the Common partition, but this
- * is written to handle other partitions if they should enter the schema.
- *
- * @param {Object} declaration - The parsed declaration
- * @param {Strint} classToFetch - The name of the class (DNS, VLAN, etc)
- * @param {function} cb - Function to execute for each object. Will be called with 2 parameters
- *                        tenant and object declaration. Object declaration is the declaration
- *                        for just the object in question, not the whole declaration
- */
-function forEach(declaration, classToFetch, cb) {
-    const tenantNames = Object.keys(declaration);
-    tenantNames.forEach((tenantName) => {
-        const tenant = declaration[tenantName];
-        const classNames = Object.keys(tenant);
-        classNames.forEach((className) => {
-            if (className === classToFetch) {
-                const classObject = tenant[className];
-                if (typeof classObject === 'object') {
-                    const containerNames = Object.keys(classObject);
-                    containerNames.forEach((containerName) => {
-                        cb(tenantName, classObject[containerName]);
-                    });
-                } else {
-                    cb(tenantName, classObject);
-                }
-            }
-        });
     });
 }